refactor: move LogSuccessMessages to helpers package
