+++ internal/helpers/logger_test.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/indaco/tempo/internal/logger"
 	"github.com/indaco/tempo/internal/testhelpers"
+	"github.com/indaco/tempo/internal/testutils"
 )
 
 func TestEnableLoggerIndentation(t *testing.T) {
@@ -58,5 +59,42 @@ func TestResetLogger(t *testing.T) {
 	// After reset, the output should not have the extra indentation.
 	if strings.HasPrefix(outputReset, "  ") {
 		t.Errorf("Expected output not to be indented after reset, got: %q", outputReset)
+	}
+}
+
+func TestLogSuccessMessages(t *testing.T) {
+	tests := []struct {
+		name       string
+		entityType string
+		expected   string
+	}{
+		{
+			name:       "Component",
+			entityType: "component",
+			expected:   "✔ Templates for the component and assets (CSS and JS) have been created",
+		},
+		{
+			name:       "Component Variant",
+			entityType: "component-variant",
+			expected:   "✔ Templates for the component variant and assets (CSS) have been created",
+		},
+		{
+			name:       "Default Case",
+			entityType: "unknown",
+			expected:   "✔ Templates and assets have been created",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			mockLogger := &testutils.MockLogger{}
+			cfg := testutils.SetupConfig(t.TempDir(), nil)
+
+			LogSuccessMessages(tt.entityType, cfg, mockLogger)
+
+			if mockLogger.Logs[0] != tt.expected {
+				t.Errorf("expected %q, got %q", tt.expected, mockLogger.Logs[0])
+			}
+		})
 	}
 }
